Persist last-file-saved in a cross session manner using localStorage
