+++ src/types.ts
@@ -52,5 +52,7 @@ export type SessionType = {
   id: string;
   path: string;
   cells: CellType[];
-  saveFile?: string;
+  // A unique identifier for this Srcbook which persists cross sessions
+  // It is used to store the files on the disk under ~/.srcbook/<dirName>
+  dirName: string;
 };
